Revert "feat(web): notification settings (#126)"
This reverts commit 4eee03c7e82ada8b1c5b5aef51bf19264000f0e6.
