+++ protobuf/proto/crux.proto
@@ -826,14 +826,6 @@ enum NotificationType {
   TEAMS = 3;
 }
 
-enum NotificationEventType {
-  UNKNOWN_NOTIFICATION_EVENT_TYPE = 0;
-  DEPLOYMENT_CREATED = 1;
-  VERSION_CREATED = 2;
-  NODE_ADDED = 3;
-  USER_INVITED = 4;
-}
-
 message CreateNotificationRequest {
   string accessedBy = 2;
 
@@ -841,8 +833,6 @@ message CreateNotificationRequest {
   string url = 101;
   NotificationType type = 102;
   bool active = 103;
-
-  repeated NotificationEventType events = 1000;
 }
 
 message CreateNotificationResponse {
@@ -859,8 +849,6 @@ message UpdateNotificationRequest {
   string url = 101;
   NotificationType type = 102;
   bool active = 103;
-  
-  repeated NotificationEventType events = 1000; 
 }
 
 message NotificationDetailsResponse {
@@ -871,8 +859,6 @@ message NotificationDetailsResponse {
   string url = 101;
   NotificationType type = 102;
   bool active = 103;
-
-  repeated NotificationEventType events = 1000; 
 }
 
 message NotificationResponse {
@@ -883,8 +869,6 @@ message NotificationResponse {
   string url = 101;
   NotificationType type = 103;
   bool active = 104;
-
-  repeated NotificationEventType events = 1000; 
 }
 
 message NotificationListResponse { repeated NotificationResponse data = 1000; }
