+++ web/crux-ui/src/components/notifications/edit-notification-card.tsx
@@ -1,3 +1,5 @@
+import { WEBOOK_TEST_DELAY } from '@app/const'
+import { DyoButton } from '@app/elements/dyo-button'
 import { DyoCard } from '@app/elements/dyo-card'
 import DyoChips from '@app/elements/dyo-chips'
 import { DyoHeading } from '@app/elements/dyo-heading'
@@ -5,21 +7,21 @@ import { DyoInput } from '@app/elements/dyo-input'
 import { DyoLabel } from '@app/elements/dyo-label'
 import { DyoSwitch } from '@app/elements/dyo-switch'
 import { defaultApiErrorHandler } from '@app/errors'
+import { useThrottling } from '@app/hooks/use-throttleing'
 import {
   CreateNotification,
   NotificationDetails,
   NotificationType,
-  NOTIFICATION_EVENT_VALUES,
   NOTIFICATION_TYPE_VALUES,
   UpdateNotification,
 } from '@app/models'
-import { API_NOTIFICATIONS, notificationApiUrl } from '@app/routes'
+import { API_NOTIFICATIONS, notificationApiHookUrl, notificationApiUrl } from '@app/routes'
 import { sendForm } from '@app/utils'
 import { notificationSchema } from '@app/validation'
 import { useFormik } from 'formik'
 import useTranslation from 'next-translate/useTranslation'
 import { MutableRefObject, useState } from 'react'
-import { NotificationEventList } from './notification-event-list'
+import toast from 'react-hot-toast'
 
 interface EditNotificationCardProps {
   notification?: NotificationDetails
@@ -39,9 +41,10 @@ const EditNotificationCard = (props: EditNotificationCardProps) => {
       type: 'discord',
       url: '',
       creator: '',
-      events: [...NOTIFICATION_EVENT_VALUES],
     },
   )
+
+  const throttle = useThrottling(WEBOOK_TEST_DELAY)
 
   const editMode = !!notification.id
 
@@ -79,6 +82,18 @@ const EditNotificationCard = (props: EditNotificationCardProps) => {
     props.submitRef.current = formik.submitForm
   }
 
+  const onTestHook = async () => {
+    if (!editMode) {
+      return
+    }
+
+    const res = await fetch(notificationApiHookUrl(notification.id), {
+      method: 'POST',
+    })
+
+    res.ok ? toast.success(t('hook.success')) : toast.error(t('hook.error'))
+  }
+
   return (
     <DyoCard className={props.className}>
       <DyoHeading element="h4" className="text-lg text-bright">
@@ -99,7 +114,19 @@ const EditNotificationCard = (props: EditNotificationCardProps) => {
             value={formik.values.name}
             message={formik.errors.name}
           />
+          <DyoInput
+            className="max-w-lg"
+            grow
+            name="url"
+            type="url"
+            label={t('url')}
+            onChange={formik.handleChange}
+            value={formik.values.url}
+            message={formik.errors.url}
+          />
+        </div>
 
+        <div className="flex flex-col">
           <div className="flex flex-col flex-wrap">
             <DyoLabel className="mt-8 mb-2.5">{t('notificationType')}</DyoLabel>
 
@@ -111,33 +138,19 @@ const EditNotificationCard = (props: EditNotificationCardProps) => {
             />
           </div>
 
-          <DyoInput
-            className="max-w-lg"
-            grow
-            name="url"
-            type="url"
-            label={t('url')}
-            onChange={formik.handleChange}
-            value={formik.values.url}
-            message={formik.errors.url}
-          />
-
           <div className="flex flex-row justify-between mt-8 items-end">
             <div className="flex flex-col">
               <DyoLabel className="mb-2.5">{t('active')}</DyoLabel>
 
               <DyoSwitch fieldName="active" checked={formik.values.active} setFieldValue={formik.setFieldValue} />
             </div>
+
+            {editMode && (
+              <DyoButton type="button" className="px-4 whitespace-nowrap" onClick={() => throttle(onTestHook)}>
+                {t('hook.textWebhook')}
+              </DyoButton>
+            )}
           </div>
-        </div>
-
-        <div className="flex flex-col">
-          <DyoLabel className="mb-2.5">{t('events')}</DyoLabel>
-
-          <NotificationEventList
-            value={formik.values.events}
-            onChanged={value => formik.setFieldValue('events', value, false)}
-          />
         </div>
       </form>
     </DyoCard>
