+++ web/crux-ui/src/pages/notifications/[notificationId].tsx
@@ -4,19 +4,15 @@ import NotificationCard from '@app/components/notifications/notification-card'
 import { BreadcrumbLink } from '@app/components/shared/breadcrumb'
 import PageHeading from '@app/components/shared/page-heading'
 import { DetailsPageMenu } from '@app/components/shared/page-menu'
-import { WEBOOK_TEST_DELAY } from '@app/const'
-import { DyoButton } from '@app/elements/dyo-button'
 import { defaultApiErrorHandler } from '@app/errors'
-import { useThrottling } from '@app/hooks/use-throttleing'
 import { NotificationDetails } from '@app/models'
-import { notificationApiHookUrl, notificationApiUrl, notificationUrl, ROUTE_NOTIFICATIONS } from '@app/routes'
+import { notificationApiUrl, notificationUrl, ROUTE_NOTIFICATIONS } from '@app/routes'
 import { withContextAuthorization } from '@app/utils'
 import { cruxFromContext } from '@server/crux/crux'
 import { NextPageContext } from 'next'
 import useTranslation from 'next-translate/useTranslation'
 import { useRouter } from 'next/router'
 import { useRef, useState } from 'react'
-import toast from 'react-hot-toast'
 
 interface NotificationDetailsPageProps {
   notification: NotificationDetails
@@ -30,8 +26,6 @@ const NotificationDetailsPage = (props: NotificationDetailsPageProps) => {
   const [editing, setEditing] = useState(false)
   const submitRef = useRef<() => Promise<void>>()
   const handleApiError = defaultApiErrorHandler(t)
-
-  const throttle = useThrottling(WEBOOK_TEST_DELAY)
 
   const onDelete = async () => {
     const res = await fetch(notificationApiUrl(notification.id), {
@@ -56,14 +50,6 @@ const NotificationDetailsPage = (props: NotificationDetailsPageProps) => {
     setNotification(notification as NotificationDetails)
   }
 
-  const onTestHook = async () => {
-    const res = await fetch(notificationApiHookUrl(notification.id), {
-      method: 'POST',
-    })
-
-    res.ok ? toast.success(t('hook.success')) : toast.error(t('hook.error'))
-  }
-
   return (
     <Layout title={t('notificationsName', { name: notification.name })}>
       <PageHeading
@@ -75,11 +61,6 @@ const NotificationDetailsPage = (props: NotificationDetailsPageProps) => {
           },
         ]}
       >
-        {!editing && (
-          <DyoButton type="button" className="px-4 mr-4 whitespace-nowrap" onClick={() => throttle(onTestHook)}>
-            {t('hook.textWebhook')}
-          </DyoButton>
-        )}
         <DetailsPageMenu
           onDelete={onDelete}
           editing={editing}
