+++ web/crux/prisma/schema.prisma
@@ -253,19 +253,7 @@ model Notification {
   
   team        Team                  @relation(fields: [teamId], references: [id], onDelete: Cascade)
 
-  events      NotificationEvent[]
-
   @@unique([name, teamId])
-}
-
-model NotificationEvent {
-  id              String                    @id @default(uuid()) @db.Uuid
-  event           NotificationEventTypeEnum
-  notificationId  String                    @db.Uuid
-
-  notification    Notification              @relation(fields: [notificationId], references: [id], onDelete: Cascade)
-
-  @@unique([event, notificationId])
 }
 
 enum ProductTypeEnum {
@@ -320,11 +308,4 @@ enum NotificationTypeEnum {
   discord
   slack
   teams
-}
-
-enum NotificationEventTypeEnum {
-  deploymentCreated
-  versionCreated
-  nodeAdded
-  userInvited
 }
