+++ web/crux/src/app/notification/notification.service.ts
@@ -42,15 +42,6 @@ export class NotificationService {
         type: this.mapper.typeToDb(request.type),
         createdBy: request.accessedBy,
         active: !!request.active,
-        events: {
-          createMany: {
-            data: (request.events ?? []).map(it => {
-              return {
-                event: this.mapper.eventTypeToDb(it),
-              }
-            }),
-          },
-        },
       },
     })
 
@@ -60,16 +51,6 @@ export class NotificationService {
   }
 
   async updateNotification(request: UpdateNotificationRequest): Promise<UpdateEntityResponse> {
-    const notificationEvents = await this.prisma.notificationEvent.findMany({
-      where: {
-        notificationId: request.id,
-      },
-    })
-
-    const eventsDbMapped = (request.events ?? []).map(ev => this.mapper.eventTypeToDb(ev))
-    const newEvents = eventsDbMapped.filter(event => !notificationEvents.find(it => it.event == event))
-    const deleteEvents = notificationEvents.filter(event => !eventsDbMapped.find(it => event.event == it))
-
     const notification = await this.prisma.notification.update({
       where: {
         id: request.id,
@@ -81,20 +62,6 @@ export class NotificationService {
         type: this.mapper.typeToDb(request.type),
         updatedBy: request.accessedBy,
         updatedAt: new Date(),
-        events: {
-          deleteMany: {
-            id: {
-              in: deleteEvents.map(ev => ev.id),
-            },
-          },
-          createMany: {
-            data: newEvents.map(it => {
-              return {
-                event: it,
-              }
-            }),
-          },
-        },
       },
     })
 
@@ -121,9 +88,6 @@ export class NotificationService {
           },
         },
       },
-      include: {
-        events: true,
-      },
     })
 
     const userIds = notifications.map(r => r.createdBy)
@@ -138,9 +102,6 @@ export class NotificationService {
     const notification = await this.prisma.notification.findUniqueOrThrow({
       where: {
         id: request.id,
-      },
-      include: {
-        events: true,
       },
     })
