+++ web/crux/src/grpc/protobuf/proto/crux.ts
@@ -409,56 +409,6 @@ export function notificationTypeToJSON(object: NotificationType): string {
   }
 }
 
-export enum NotificationEventType {
-  UNKNOWN_NOTIFICATION_EVENT_TYPE = 0,
-  DEPLOYMENT_CREATED = 1,
-  VERSION_CREATED = 2,
-  NODE_ADDED = 3,
-  USER_INVITED = 4,
-  UNRECOGNIZED = -1,
-}
-
-export function notificationEventTypeFromJSON(object: any): NotificationEventType {
-  switch (object) {
-    case 0:
-    case 'UNKNOWN_NOTIFICATION_EVENT_TYPE':
-      return NotificationEventType.UNKNOWN_NOTIFICATION_EVENT_TYPE
-    case 1:
-    case 'DEPLOYMENT_CREATED':
-      return NotificationEventType.DEPLOYMENT_CREATED
-    case 2:
-    case 'VERSION_CREATED':
-      return NotificationEventType.VERSION_CREATED
-    case 3:
-    case 'NODE_ADDED':
-      return NotificationEventType.NODE_ADDED
-    case 4:
-    case 'USER_INVITED':
-      return NotificationEventType.USER_INVITED
-    case -1:
-    case 'UNRECOGNIZED':
-    default:
-      return NotificationEventType.UNRECOGNIZED
-  }
-}
-
-export function notificationEventTypeToJSON(object: NotificationEventType): string {
-  switch (object) {
-    case NotificationEventType.UNKNOWN_NOTIFICATION_EVENT_TYPE:
-      return 'UNKNOWN_NOTIFICATION_EVENT_TYPE'
-    case NotificationEventType.DEPLOYMENT_CREATED:
-      return 'DEPLOYMENT_CREATED'
-    case NotificationEventType.VERSION_CREATED:
-      return 'VERSION_CREATED'
-    case NotificationEventType.NODE_ADDED:
-      return 'NODE_ADDED'
-    case NotificationEventType.USER_INVITED:
-      return 'USER_INVITED'
-    default:
-      return 'UNKNOWN'
-  }
-}
-
 export enum ServiceStatus {
   UNKNOWN_SERVICE_STATUS = 0,
   UNAVAILABLE = 1,
@@ -1091,7 +1041,6 @@ export interface CreateNotificationRequest {
   url: string
   type: NotificationType
   active: boolean
-  events: NotificationEventType[]
 }
 
 export interface CreateNotificationResponse {
@@ -1106,7 +1055,6 @@ export interface UpdateNotificationRequest {
   url: string
   type: NotificationType
   active: boolean
-  events: NotificationEventType[]
 }
 
 export interface NotificationDetailsResponse {
@@ -1116,7 +1064,6 @@ export interface NotificationDetailsResponse {
   url: string
   type: NotificationType
   active: boolean
-  events: NotificationEventType[]
 }
 
 export interface NotificationResponse {
@@ -1126,7 +1073,6 @@ export interface NotificationResponse {
   url: string
   type: NotificationType
   active: boolean
-  events: NotificationEventType[]
 }
 
 export interface NotificationListResponse {
@@ -3408,7 +3354,6 @@ const baseCreateNotificationRequest: object = {
   url: '',
   type: 0,
   active: false,
-  events: 0,
 }
 
 export const CreateNotificationRequest = {
@@ -3421,7 +3366,6 @@ export const CreateNotificationRequest = {
     message.url = object.url !== undefined && object.url !== null ? String(object.url) : ''
     message.type = object.type !== undefined && object.type !== null ? notificationTypeFromJSON(object.type) : 0
     message.active = object.active !== undefined && object.active !== null ? Boolean(object.active) : false
-    message.events = (object.events ?? []).map((e: any) => notificationEventTypeFromJSON(e))
     return message
   },
 
@@ -3432,11 +3376,6 @@ export const CreateNotificationRequest = {
     message.url !== undefined && (obj.url = message.url)
     message.type !== undefined && (obj.type = notificationTypeToJSON(message.type))
     message.active !== undefined && (obj.active = message.active)
-    if (message.events) {
-      obj.events = message.events.map(e => notificationEventTypeToJSON(e))
-    } else {
-      obj.events = []
-    }
     return obj
   },
 }
@@ -3468,7 +3407,6 @@ const baseUpdateNotificationRequest: object = {
   url: '',
   type: 0,
   active: false,
-  events: 0,
 }
 
 export const UpdateNotificationRequest = {
@@ -3482,7 +3420,6 @@ export const UpdateNotificationRequest = {
     message.url = object.url !== undefined && object.url !== null ? String(object.url) : ''
     message.type = object.type !== undefined && object.type !== null ? notificationTypeFromJSON(object.type) : 0
     message.active = object.active !== undefined && object.active !== null ? Boolean(object.active) : false
-    message.events = (object.events ?? []).map((e: any) => notificationEventTypeFromJSON(e))
     return message
   },
 
@@ -3494,11 +3431,6 @@ export const UpdateNotificationRequest = {
     message.url !== undefined && (obj.url = message.url)
     message.type !== undefined && (obj.type = notificationTypeToJSON(message.type))
     message.active !== undefined && (obj.active = message.active)
-    if (message.events) {
-      obj.events = message.events.map(e => notificationEventTypeToJSON(e))
-    } else {
-      obj.events = []
-    }
     return obj
   },
 }
@@ -3509,7 +3441,6 @@ const baseNotificationDetailsResponse: object = {
   url: '',
   type: 0,
   active: false,
-  events: 0,
 }
 
 export const NotificationDetailsResponse = {
@@ -3524,7 +3455,6 @@ export const NotificationDetailsResponse = {
     message.url = object.url !== undefined && object.url !== null ? String(object.url) : ''
     message.type = object.type !== undefined && object.type !== null ? notificationTypeFromJSON(object.type) : 0
     message.active = object.active !== undefined && object.active !== null ? Boolean(object.active) : false
-    message.events = (object.events ?? []).map((e: any) => notificationEventTypeFromJSON(e))
     return message
   },
 
@@ -3536,11 +3466,6 @@ export const NotificationDetailsResponse = {
     message.url !== undefined && (obj.url = message.url)
     message.type !== undefined && (obj.type = notificationTypeToJSON(message.type))
     message.active !== undefined && (obj.active = message.active)
-    if (message.events) {
-      obj.events = message.events.map(e => notificationEventTypeToJSON(e))
-    } else {
-      obj.events = []
-    }
     return obj
   },
 }
@@ -3551,7 +3476,6 @@ const baseNotificationResponse: object = {
   url: '',
   type: 0,
   active: false,
-  events: 0,
 }
 
 export const NotificationResponse = {
@@ -3564,7 +3488,6 @@ export const NotificationResponse = {
     message.url = object.url !== undefined && object.url !== null ? String(object.url) : ''
     message.type = object.type !== undefined && object.type !== null ? notificationTypeFromJSON(object.type) : 0
     message.active = object.active !== undefined && object.active !== null ? Boolean(object.active) : false
-    message.events = (object.events ?? []).map((e: any) => notificationEventTypeFromJSON(e))
     return message
   },
 
@@ -3576,11 +3499,6 @@ export const NotificationResponse = {
     message.url !== undefined && (obj.url = message.url)
     message.type !== undefined && (obj.type = notificationTypeToJSON(message.type))
     message.active !== undefined && (obj.active = message.active)
-    if (message.events) {
-      obj.events = message.events.map(e => notificationEventTypeToJSON(e))
-    } else {
-      obj.events = []
-    }
     return obj
   },
 }
