+++ web/crux/src/services/domain.notification.service.ts
@@ -1,12 +1,11 @@
 import { nameOrEmailOfIdentity } from './../shared/model'
 import { HttpService } from '@nestjs/axios'
 import { Injectable, Logger } from '@nestjs/common'
-import { NotificationEventTypeEnum, NotificationTypeEnum } from '@prisma/client'
+import { NotificationTypeEnum } from '@prisma/client'
 import { lastValueFrom } from 'rxjs'
-import { getTemplate, NotificationMessageType, NotificationTemplate } from 'src/domain/notification-templates'
+import { getTemplate, NotificationTemplate } from 'src/domain/notification-templates'
 import { KratosService } from './kratos.service'
 import { PrismaService } from './prisma.service'
-import { InvalidArgumentException } from 'src/exception/errors'
 
 @Injectable()
 export class DomainNotificationService {
@@ -15,8 +14,6 @@ export class DomainNotificationService {
   constructor(private prisma: PrismaService, private httpService: HttpService, private kratos: KratosService) {}
 
   async sendNotification(template: NotificationTemplate): Promise<void> {
-    const eventType = this.getMessageEventFromType(template.messageType)
-
     const userOnTeam = await this.prisma.usersOnTeams.findFirst({
       where: {
         userId: template.identityId,
@@ -36,17 +33,7 @@ export class DomainNotificationService {
       })
 
       notifications.forEach(async notification => {
-        const count = await this.prisma.notificationEvent.count({
-          where: {
-            notificationId: notification.id,
-            event: eventType,
-          },
-          take: 1,
-        })
-
-        if (count > 0) {
-          await this.send(notification.url, notification.type, template)
-        }
+        await this.send(notification.url, notification.type, template)
       })
     }
   }
@@ -60,27 +47,6 @@ export class DomainNotificationService {
       }
     } catch (err) {
       this.logger.error(err)
-    }
-  }
-
-  private getMessageEventFromType(messageType: NotificationMessageType): NotificationEventTypeEnum {
-    switch (messageType) {
-      case 'node':
-        return NotificationEventTypeEnum.nodeAdded
-      case 'version':
-        return NotificationEventTypeEnum.versionCreated
-      case 'invite':
-        return NotificationEventTypeEnum.userInvited
-      case 'failedDeploy':
-        return NotificationEventTypeEnum.deploymentCreated
-      case 'successfulDeploy':
-        return NotificationEventTypeEnum.deploymentCreated
-      default:
-        throw new InvalidArgumentException({
-          property: 'messageType',
-          value: messageType,
-          message: `Unknown NotificationMessageType '${messageType}'`,
-        })
     }
   }
 }
